refactor(core): extract wasm program wrapper module
