+++ crates/tsz-core/src/api/wasm/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod code_actions;
 pub(crate) mod compiler_options;
+pub(crate) mod program;
 pub(crate) mod program_results;
